% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{:=}
\alias{.data}
\title{Objects exported from other packages}
\keyword{internal}
\value{
See documentation in each object's original package.
}
\examples{
# See examples in each object's original package
NULL
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:dyn-dots]{:=}}, \code{\link[rlang:dot-data]{.data}}}
}}

