% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_cytometry_data}
\alias{simulate_cytometry_data}
\title{Simulate Cytometry Data for FlowSet and FlowFrame Analysis}
\usage{
simulate_cytometry_data(num_cells = 100, num_features = 10, num_flowframes = 5)
}
\arguments{
\item{num_cells}{An integer indicating the number of cells to simulate.}

\item{num_features}{An integer indicating how many features to simulate.}

\item{num_flowframes}{An integer indicating how many flowFrames to simulate
for the simulated flowSet.}
}
\value{
A list containing two entries: a flowFrame and a flowSet.
}
\description{
Simulate Cytometry Data for FlowSet and FlowFrame Analysis
}
\examples{
simulate_cytometry_data()
}
